requirements.txt
fastapi
uvicorn
python-jose[cryptography]
pydantic
python-dotenv
passlib
bcrypt
